requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.3.0
python-decouple==3.8
requests==2.32.4
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
